extract terraform commands
